jsonschema: #51 - split into multiple files
This change splits the encoder logic into a separate file per supported
schema.FieldValidator type. It also provides some minor clenaups to the
documentation and adds some FIXME statements.

Functions / type defintions are declared in the order in which they are
called, which should make the code "read like an article from top to
bottom".
