+++ schema/encoding/doc.go
@@ -1,0 +1,2 @@
+// Package encoding is the intended hierarchy location for encoding Schema to other formats.
+package encoding
