--- schema/encoding/encoding.go
@@ -1,3 +1,0 @@
-package encoding
-
-// Package encoding is the intended hierarchy location for Encoders for Schema to other formats.
