+++ schema/encoding/jsonschema/doc.go
@@ -1,0 +1,4 @@
+// Package jsonschema provides JSON Schema Draft 4 encoding support for schema.Schema. Note that the current
+// implementation is incomlete, and not all FieldValidator types are yet supported. Custom validators are also not
+// supported at the momenet.
+package jsonschema
