+++ schema/encoding/jsonschema/object.go
@@ -1,0 +1,15 @@
+package jsonschema
+
+import (
+	"io"
+
+	"github.com/rs/rest-layer/schema"
+)
+
+func encodeObject(w io.Writer, v *schema.Object) error {
+	ew := errWriter{w: w}
+	if ew.err == nil && v.Schema != nil {
+		ew.err = schemaToJSONSchema(w, v.Schema)
+	}
+	return ew.err
+}
